Subject: function that concatenates two strings using n bytes of second string, if n >= 2nd string len use 2nd string

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -9,7 +9,7 @@
  */
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
-	size_t len1 = 0, len2 = 0, total_len = 0, i = 0, j = 0;
+	unsigned int len1 = 0, len2 = 0, total_len = 0, i = 0, j = 0;
 	char *newString;
 
 	while (s1[len1] != '\0')
